Execute job requests in chunks
For one example project, this part started breaking down with many
requests to the separate job endpoints. My suspicion that it had to do
with the number of requests seemed to be confirmed when I copy&pasted
some code that splits this up into chunks of 50, and it started working
again.