Refactoring: Make GitlabClient dumber
Pulling out some logic how to deal with the change data into a new
component, CdChangeService. Leaving the CdChangeReader to deal with some
simple things like just loading info on branches, tags and commits.

In preparation for getting a better implementation for release branches,
it's easier to reason about when the GitlabAPI and dealing with the
information are separated from each other