Focus on pipeline failure rates
Looking at overall job failure rates doesn't make as much sense - there
are potentially a lot of them, depending on how much the pipelines are
broken into separate steps. The more fine granular a team decides to
make their pipelines, the more it dilutes the failure rate. What
actually counts is an overall pipeline and if it gets stuck.