Remember original name of a branch
Mainly for debugging purposes - when using the GitRepoClient, the
branches have full-blown names like /refs/remote/..., which gets in the
way of readability, and is also a hassle to pass on as branch name
parameters. So CDReference.name is a short, humanized version of a branch
name, CDReference.originalName will contain that full path name.